requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.5
curl_cffi==0.13.0
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
Farama-Notifications==0.0.4
filelock==3.21.2
fonttools==4.61.1
frozendict==2.4.7
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.63.0
groq==1.0.0
gymnasium==1.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multitasking==0.0.12
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
openai==2.20.0
packaging==26.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.1
platformdirs==4.7.0
plotly==6.5.2
pluggy==1.6.0

protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
schedule==1.2.2
setuptools==82.0.0
Shimmy==2.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
stable_baselines3==2.7.1
streamlit==1.54.0
sympy==1.14.0
ta==0.11.0
tenacity==9.1.4
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
yfinance==1.1.0
